refactor(path): Template Method Pattern을 사용하여 PathFinder 구현체의 중복 로직 빼냄
